image-source: Merge changes into main slideshow
