+++ plugins/image-source/CMakeLists.txt
@@ -8,8 +8,7 @@ endif()
 set(image-source_SOURCES
 	image-source.c
 	color-source.c
-	obs-slideshow.c
-	obs-slideshow2.c)
+	obs-slideshow.c)
 
 add_library(image-source MODULE
 	${image-source_SOURCES})
