+++ plugins/image-source/data/locale/de-DE.ini
@@ -29,33 +29,6 @@ SlideShow.NextSlide="Nächstes Bild"
 SlideShow.PreviousSlide="Vorheriges Bild"
 SlideShow.HideWhenDone="Verbergen, wenn Diashow vorbei"
 
-SlideShow2="Diashow"
-SlideShow2.TransitionSpeed="Übergangsgeschwindigkeit (Millisekunden)"
-SlideShow2.SlideTime="Zeit zwischen Bildern (Millisekunden)"
-SlideShow2.Files="Bilddateien"
-SlideShow2.CustomSize="Rahmengröße/Seitenverhältnis"
-SlideShow2.CustomSize.Auto="Automatisch"
-SlideShow2.Randomize="Zufällige Wiedergabe"
-SlideShow2.Loop="Endlosschleife"
-SlideShow2.Transition="Übergang"
-SlideShow2.Transition.Cut="Schnitt"
-SlideShow2.Transition.Fade="Überblenden"
-SlideShow2.Transition.Swipe="Swipe"
-SlideShow2.Transition.Slide="Slide"
-SlideShow2.PlaybackBehavior="Sichtbarkeitsverhalten"
-SlideShow2.PlaybackBehavior.StopRestart="Anhalten, wenn nicht sichtbar; neustarten, wenn sichtbar"
-SlideShow2.PlaybackBehavior.PauseUnpause="Pausieren, wenn nicht sichtbar; fortsetzen, wenn sichtbar"
-SlideShow2.PlaybackBehavior.AlwaysPlay="Immer abspielen, auch wenn nicht sichtbar"
-SlideShow2.SlideMode="Diashowmodus"
-SlideShow2.SlideMode.Auto="Automatisch"
-SlideShow2.SlideMode.Manual="Manuell (Hotkeys für die Steuerung der Diashow verwenden)"
-SlideShow2.PlayPause="Abspielen/pausieren"
-SlideShow2.Restart="Neustarten"
-SlideShow2.Stop="Stoppen"
-SlideShow2.NextSlide="Nächstes Bild"
-SlideShow2.PreviousSlide="Vorheriges Bild"
-SlideShow2.HideWhenDone="Verbergen, wenn Diashow vorbei"
-
 ColorSource="Farbquelle"
 ColorSource.Color="Farbe"
 ColorSource.Width="Breite"
