+++ plugins/image-source/data/locale/en-US.ini
@@ -29,33 +29,6 @@ SlideShow.NextSlide="Next Slide"
 SlideShow.PreviousSlide="Previous Slide"
 SlideShow.HideWhenDone="Hide when slideshow is done"
 
-SlideShow2="Image Slide Show"
-SlideShow2.TransitionSpeed="Transition Speed (milliseconds)"
-SlideShow2.SlideTime="Time Between Slides (milliseconds)"
-SlideShow2.Files="Image Files"
-SlideShow2.CustomSize="Bounding Size/Aspect Ratio"
-SlideShow2.CustomSize.Auto="Automatic"
-SlideShow2.Randomize="Randomize Playback"
-SlideShow2.Loop="Loop"
-SlideShow2.Transition="Transition"
-SlideShow2.Transition.Cut="Cut"
-SlideShow2.Transition.Fade="Fade"
-SlideShow2.Transition.Swipe="Swipe"
-SlideShow2.Transition.Slide="Slide"
-SlideShow2.PlaybackBehavior="Visibility Behavior"
-SlideShow2.PlaybackBehavior.StopRestart="Stop when not visible, restart when visible"
-SlideShow2.PlaybackBehavior.PauseUnpause="Pause when not visible, unpause when visible"
-SlideShow2.PlaybackBehavior.AlwaysPlay="Always play even when not visible"
-SlideShow2.SlideMode="Slide Mode"
-SlideShow2.SlideMode.Auto="Automatic"
-SlideShow2.SlideMode.Manual="Manual (Use hotkeys to control slideshow)"
-SlideShow2.PlayPause="Play/Pause"
-SlideShow2.Restart="Restart"
-SlideShow2.Stop="Stop"
-SlideShow2.NextSlide="Next Slide"
-SlideShow2.PreviousSlide="Previous Slide"
-SlideShow2.HideWhenDone="Hide when slideshow is done"
-
 ColorSource="Color Source"
 ColorSource.Color="Color"
 ColorSource.Width="Width"
